+++ ipro-intern-project/backend/run_test.py
@@ -1,0 +1,14 @@
+from src import session
+from src import data_types
+import unittest
+import faker
+
+s = session()
+
+class TestTables(unittest.testCase):
+    def setUp(self):
+        pass
+
+    def tearDown(self):
+        pass
+
