+++ ipro-intern-project/backend/src/start.bat
@@ -1,0 +1,1 @@
+uvicorn main:app --reload
