--- ipro-intern-project/backend/start.bat
@@ -1,1 +1,0 @@
-uvicorn main:app --reload
